+++ R/tsibble2ts.R
@@ -22,16 +22,13 @@
 as.ts.tbl_ts <- function(x, value, frequency = NULL, fill = NA, ...) {
   value <- enquo(value)
   key_vars <- key(x)
-  if (is_empty(key_vars)) {
-    return(finalise_ts(x, index = index(x), frequency = frequency))
-  }
   if (any(is_nest(key_vars)) || length(key_vars) > 1) {
     abort("`as.ts()` can't deal with nested or crossed keys.")
   }
   mvars <- measured_vars(x)
   str_val <- paste_comma(surround(mvars, "`"))
   if (quo_is_missing(value)) {
-    if (is_false(has_length(mvars, 1))) {
+    if (is_false(has_length(mvars, 1) || is_empty(key_vars))) {
       abort(sprintf("Can't determine the `value`: %s.", str_val))
     }
     value_var <- mvars
@@ -42,11 +39,14 @@ as.ts.tbl_ts <- function(x, value, frequency = NULL, fill = NA, ...) {
     }
   }
   idx <- index(x)
-  key <- key_vars(x)
-  mat_ts <- x %>% 
+  tsbl_sel <- x %>% 
     arrange(!!! key_vars, !! idx) %>% 
-    select(!! idx, !!! key_vars, !! value_var) %>% 
-    spread(key = !! key, value = !! value_var, fill = fill)
+    select(!! idx, !!! key_vars, !! value_var)
+  if (is_empty(key_vars)) {
+    return(finalise_ts(tsbl_sel, index = index(x), frequency = frequency))
+  }
+  mat_ts <- tsbl_sel %>% 
+    spread(key = !! key_vars[[1]], value = !! value_var, fill = fill)
   finalise_ts(mat_ts, index = idx, frequency = frequency)
 }